converted gatsby 2.X image and richtext query syntax
