+++ src/fragments/NodeImageFields.ts
@@ -1,0 +1,12 @@
+import { graphql } from "gatsby";
+
+export const NodeImageFields = graphql`
+  fragment NodeImageFields on ContentfulAsset {
+    title
+    file {
+      url
+      contentType
+    }
+    description
+  }
+`;
