--- src/node/post.js
@@ -1,8 +1,0 @@
-import React from "react";
-import Layout from "../components/layout/MainLayout";
-
-export default ({ pageContext }) => (
-  <Layout>
-    <h2>Post Here</h2> {console.log(pageContext)}
-  </Layout>
-);
